added pages, reorganized to accomodate pages
