+++ Portfolio/package.json
@@ -11,8 +11,11 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "dom": "^0.0.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.28.0",
+    "router": "^1.3.8",
     "uuid": "^11.0.3"
   },
   "devDependencies": {
