+++ Portfolio/src/main.jsx
@@ -2,9 +2,27 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.jsx'
+import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+import Home from './pages/Home.jsx';
+import ContactMe from './pages/ContactMe.jsx';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <App />,
+    children: [
+      {
+        index: true,
+        element: <Home />,
+      },
+      {
+        path: '/ContactMe',
+        element: <ContactMe />,
+      }
+    ],
+  },
+]);
 
 createRoot(document.getElementById('root')).render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
-)
+  <RouterProvider router={router} />
+);
